打赏礼物扣除对应的零钱并提醒若余额不足去充值

+++ app/src/main/java/cn/ucai/live/ui/activity/ChangeActivity.java
@@ -27,7 +27,7 @@ public class ChangeActivity extends AppCompatActivity {
     @BindView(R.id.target_layout)
     LinearLayout targetLayout;
     View loadingView;
-    int change;
+    long change;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
